nits

+++ src/data.rs
@@ -205,16 +205,16 @@ pub struct ProgLoc {
 impl ProgLoc {
     pub fn print_location(loc: &ProgLoc, lines: &[&str], itr: &usize) -> bool {
         if loc.line == 0
-            || (loc.line as usize) > lines.len()
-            || loc.char_range.0 >= lines[(loc.line - 1) as usize].len()
-            || loc.char_range.1 > lines[(loc.line - 1) as usize].len()
+            || loc.line > lines.len()
+            || loc.char_range.0 >= lines[loc.line - 1].len()
+            || loc.char_range.1 > lines[loc.line - 1].len()
             || loc.char_range.0 >= loc.char_range.1
         {
             // println!("Invalid location: {:?}", loc);
             return false;
         }
 
-        let line_text = lines[(loc.line - 1) as usize];
+        let line_text = lines[loc.line - 1];
 
         println!(
             "{} {} {} {} {} {}",
